package/monit: bump version to 5.43.3
Patch 0002 needs to be slightly refreshed to apply on this new
version.

Version 5.34.3
==============

Fixed: If the ping statement did not explicitly specify an outgoing
address but a previous ping statement did, the same address was shared
by both statements.

Fixed: Monit may crash upon stopping if the ping statement is used in
conjunction with the address option.

Fixed: If a directory is set in the 'allow' option of the 'set httpd'
statement, instead of file or string, Monit hangs on start.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Signed-off-by: Julien Olivain <ju.o@free.fr>