package/mpdecimal: bump to version 4.0.0
While 2.5.1 -> 4.0.0 may seems like a big jump with lots of versions
skipped, in fact 4.0.0 is the immediate next release after 2.5.1.

Changelog at:
https://www.bytereef.org/mpdecimal/changelog.html

Build tested successfully with test-pkg, including with the one
reverse dependency of mpdecimal: the python3 decimal module:
https://gitlab.com/tpetazzoni/buildroot/-/pipelines/1604635008.

The license file was renamed from LICENSE.txt to COPYRIGHT.txt. The
license is still BSD-2-Clause, but the formatting changed: the license
text is no longer inside a C-style comment, but just plain text, and
the copyright years were updated.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Signed-off-by: Julien Olivain <ju.o@free.fr>