package/libinput: bump to version 1.27
From
https://lists.freedesktop.org/archives/wayland-devel/2024-November/043860.html:

In terms of new features we have a "sticky" mode for drag-locking. Previously a
tap-and-drag lock would always expire after a timeout, now the button is held
logically down until a completing tap. Desktop environments are encouraged to
use this as the default as it provides a better experience for anyone with
less-than-perfect dexterity. For backwards-compatibility reasons libinput cannot
easily change its defaults without risking bugs in the callers.

For tablet pads we now support tablet pad mode groups for devices without
status LEDs as well, the previous implementation was tied to LEDs which some
devices like the XP Pen ACK05 remote don't have. Since the mode is a software
feature anyway tying it to LEDs is not necessary.

If a tablet is unknown to libwacom we now assume that it is a built-in tablet.
This matches the behavior of libwacom 2.11 but in our case the only visible
result is that the device now has the calibration configuration available.
Better to have it and not use it, as they say.

The available area on external tablets can be reduced via the new tablet "area"
configuration. Users can set a rectangle smaller than the width/height
of the tablet and input outside this rectangle will be ignored.

For packagers: the check dependency is now optional, almost all tests
can now run without it.

And then we have of course the usual collection of bug fixes and
device-specific quirks.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Signed-off-by: Julien Olivain <ju.o@free.fr>