linux: select host-uboot-tools with FIT on mips
In the latest kernel, FIT images are always generated when building a
kernel for MIPS. Note that we build the kernel with:

  make all
  make <selected-image>

so the selected image through Buildroot options doesn't matter: a FIT
image is always generated.

Therefore, in order to fix autobuilder issues, make sure
host-uboot-tools with FIT image are always selected when building the
latest kernel version. We do not select it in general as custom
versions may be different.

Fixes:

  http://autobuild.buildroot.net/results/24bd5ae44f4097c374fe43b263254a985e2989d2 (mips)
  http://autobuild.buildroot.net/results/d18b458a600f1f4bd7008d13ad2a56cb22ce8a73 (mips64el)

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Signed-off-by: Julien Olivain <ju.o@free.fr>
(cherry picked from commit 40e266ccb9aa5a9d9ddcee2457990ee2d80c5a21)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>