package/procps-ng: bump version to 4.0.5
Patch 0001 is dropped as it's upstream as of
ca004d4657d5e8b468a4552ede429be53193a3a9.

Patch 0002 is dropped as it's upstream as of
263fe04f9ed6f3f40fdb7ce249221e03dc926c9b.

Patch 0003 is dropped as it's upstream as of
2507bc475782ff5e0541d37c780dff1e293c9553.

Patch 0004 is dropped as it's upstream as of
587efb47df7ddbfda4e6abdd1e7792a2531a238f.

Patch 0005 is dropped as it's upstream as of
5acbb5dc1587d688de646d739a97251eb893bbb0.

Therefore AUTORECONF = YES can be dropped as we're no longer patching
the package.

Release notes:
https://gitlab.com/procps-ng/procps/-/releases/v4.0.5

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Signed-off-by: Julien Olivain <ju.o@free.fr>