package/igt-gpu-tools: fix build with procps-ng 4.0.5
Since the bump of procps-ng to version 4.0.5 in Buildroot commit
f7fe84508a14641617586bc481a08cb3face0477, the build of igt-gpu-tools
fails with:

../lib/igt_aux.c:1358:78: error: macro "PIDS_VAL" passed 4 arguments, but takes just 3

Indeed, the API of procps-ng has slightly changed between 4.0.4 and
4.0.5. Since it looks strange as those are minor releases the API
breakage has been reported to upstream procps-ng.

In the mean time, we still need to fix igt-gpu-tools, which the
patch added in this commit does. This patch has been submitted to
upstream igt-gpu-tools.

There are no autobuilder failures at this point, as the update to
procps-ng is quite recent, and igt-gpu-tools needing quite a lot of
dependencies, the build has a high chance of failing elsewhere
previously. However, the build issue is reproducible with the
following minimal defconfig:

 BR2_arm=y
 BR2_cortex_a9=y
 BR2_ARM_ENABLE_VFP=y
 BR2_TOOLCHAIN_EXTERNAL=y
 BR2_TOOLCHAIN_EXTERNAL_BOOTLIN=y
 BR2_INIT_NONE=y
 BR2_ROOTFS_DEVICE_CREATION_DYNAMIC_EUDEV=y
 BR2_SYSTEM_BIN_SH_NONE=y
 # BR2_PACKAGE_BUSYBOX is not set
 BR2_PACKAGE_IGT_GPU_TOOLS=y
 # BR2_TARGET_ROOTFS_TAR is not set

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Signed-off-by: Julien Olivain <ju.o@free.fr>