package/gobject-introspection: disable on Microblaze
The build of GOI on Microblaze fails as qemu-user hangs forver when
running the GOI programs. Considering how small Microblaze targets
are, the use-case for GOI is very small if not inexistant, and it's
unlikely anybody is ever going to debug this, so just disable GOI on
Microblaze.

This issue is causing timeouts in the autobuilders on a regular basis:

  http://autobuild.buildroot.net/?status=TIMEOUT&reason=gobject-introspection%

Fixes:

  http://autobuild.buildroot.net/results/f8e5ef74478c63c89e7b99fb928b97ac4518f943/

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Signed-off-by: Julien Olivain <ju.o@free.fr>
(cherry picked from commit 8548c7586a56938b2f52f5c41050441b53a457f0)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>