configs/imx6ulz_bsh_smm_m2: fix kernel hash
Commit 75320f567d "configs/imx6ulz_bsh_smm_m2: bump U-Boot to 2024.10
and Linux to 6.1.120" updated the kernel from 6.1.68 to 6.1.120 but
forgot to update its per-config kernel hash file.

The defconfig fails with the error:

    ERROR: No hash found for linux-6.1.120.tar.xz

This failure happen since commit 07149e5f30 "{linux, linux-headers}:
bump 5.{4, 10, 15}.x / 6.{1, 6, 12}.x series" which updated the
Buildroot 6.1 kernel series from 6.1.120 to 6.1.122.

This commit fixes the issue by updating the hash.

While at it, it also changes the "Locally computed" comment
by the upstream link to hashes.

Fixes:
https://gitlab.com/buildroot.org/buildroot/-/jobs/8751983243

Cc: Dario Binacchi <dario.binacchi@amarulasolutions.com>
Cc: Michael Trimarchi <michael@amarulasolutions.com>
Reviewed-by: Dario Binacchi <dario.binacchi@amarulasolutions.com>
Signed-off-by: Julien Olivain <ju.o@free.fr>