package/cross-ldd: new package
This commit introduces a new cross-ldd package which installs the
cross ldd implementation that comes from Crosstool-NG, and which can
be used like ldd, but when cross-compiling. We intend to use it as a
replacement of host-prelink-cross for our dracut package.

The cross-ldd from Crosstool-NG is in fact even recommended by the
Dracut documentation, in its docs/README.cross document:

   To support cross-compiled binaries, a different ldd variant is needed that
   works on those binaries. One such ldd script is found at
   https://gist.github.com/jerome-pouiller/c403786c1394f53f44a3b61214489e6f

This Github Gist in fact contains the xldd.in code from Crosstool-NG.

Our package downloads the full tarball of Crosstool-NG. We tried to
download only the script and the license file, but we couldn't figure
out a way of downloading them *and* have the version in the filename
to make sure that when doing an upgrade to a newer version, we
wouldn't have a clash between old/new files having the same name (and
causing conflicts with the hash verification). So in the end, we opted
for download the whole tarball.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Signed-off-by: Romain Naour <romain.naour@smile.fr>