package/dracut: use host-cross-ldd instead of host-prelink-cross
This commit reworks the dracut package to use the cross-ldd installed
by host-cross-ldd instead of prelink-rtld installed by
host-prelink-cross.

Indeed, prelink-cross is not maintained upstream, and its dependency
on host-libiberty causes problems in Buildroot with other packages.

The only change needed is that the dracut_wrapper script needs to be
generated, as we now need to know the prefix of the cross-compilation
tool (TARGET_CROSS).

Tested by running the dracut tests of the test suite.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Signed-off-by: Romain Naour <romain.naour@smile.fr>