package/libiberty: remove package
host-libiberty was only used by host-prelink-cross, which has now been
removed.

The dependency of host-gdb on host-libiberty was not because it was
needed, but because if both host-libiberty and host-gdb were built,
host-gdb should be built after host-liberty. See commit
f0a583ddc4c2cab121330284e316f54229dee492 ("package/gdb: depend on
libiberty")

Fixes:

  http://autobuild.buildroot.net/results/82d2d7d30c8e00bf6e9918c3d989f65a7330d381/ (host-binutils)
  http://autobuild.buildroot.net/results/a4674122a683e4842c1fd913da80a3e6e484d5a5/ (host-binutils-bare-metal)

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Signed-off-by: Romain Naour <romain.naour@smile.fr>