package/fio: bump version to 3.38
For change log since 3.34, see:
https://git.kernel.dk/cgit/fio/log/?h=fio-3.38

While at it, this commit also updates the homepage url in Config.in
(which now points to cgit on https) and the _SITE url (which now uses
https) to match urls published in [1]. This commit also adds the md5
hash published upstream and a comment about the pgp signature check.

Also, fio includes the <linux/nvme_ioctl.h> kernel header. See [2]
and [3]. This Kernel header was added in commit [4] and [5], in
Kernel v4.5.

This commits adds this new kernel header version requirement.

[1] https://git.kernel.dk/cgit/fio/tree/README.rst?h=fio-3.38
[2] https://git.kernel.dk/cgit/fio/tree/engines/io_uring.c?h=fio-3.38#n34
[3] https://git.kernel.dk/cgit/fio/tree/engines/nvme.h?h=fio-3.38#n12
[4] https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux.git/commit/?id=9d99a8dda154f38307d43d9c9aa504bd3703d596
[5] https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux.git/commit/?id=a9cf8284b45110a4d98aea180a89c857e53bf850

Signed-off-by: Julien Olivain <ju.o@free.fr>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>